Make metadata contain full checking handshake contents
